Split connection.ts into connection, page-state, and ref-resolver modules
